docs/requirements.txt
click==7.0
pybtex==0.22.2
sphinxcontrib-bibtex==0.4.2
colorama==0.4.1
coloredlogs==10.0
